+++ packages/ts-json-api/src/types/shared.ts
@@ -1,16 +1,36 @@
 /**
+ * A single Link.
+ */
+export type Link = string | LinkObject;
+
+/**
  * An index of Links.
  */
 export interface Links {
-    [index: string]: string | LinkObject;
+    [index: string]: Link;
 }
+
+/**
+ * Top level pagination Links.
+ */
+export interface PaginationLinks {
+    first?: Link | null;
+    last?: Link | null;
+    prev?: Link | null;
+    next?: Link | null;
+}
+
+/**
+ * The top level Links.
+ */
+export type TopLevelLinks = Links & PaginationLinks;
 
 /**
  * A Link.
  */
 export interface LinkObject {
     href: string;
-    meta: Meta;
+    meta?: Meta;
 }
 
 /**